fairness_model/requirements.txt
torch
torchvision
numpy
pillow
tqdm
scikit-learn
facenet-pytorch
opencv-python-headless
